refactor: Extract MainViewModel and update initial camera position
This commit introduces the following changes:

-   **New `MainViewModel.kt`**:
    -   Extracts the `MainViewModel` class from `MainActivity.kt` into its own file. This ViewModel is responsible for holding the selected place ID, allowing the state to survive configuration changes.
-   **`MainActivity.kt` updates**:
    -   Removes the inline `MainViewModel` class definition.
    -   Updates the `showPlaceDetailsFragment` function to launch coroutines within `lifecycleScope` and use `withContext(Dispatchers.Main)` for UI operations.
    -   Moves the `initialCamera` definition into a companion object.
    -   Modifies the `initialCamera` coordinates to point to a new location (near Boulder, CO), and updates its altitude, heading, tilt, and range.
-   **`activity_main.xml` updates**:
    -   Updates the `map3d:centerLat`, `map3d:centerLng`, `map3d:centerAlt`, `map3d:heading`, `map3d:tilt`, and `map3d:range` attributes for the `Map3DView` to reflect the new initial camera position.
