+++ PlacesUIKit3D/app/src/main/res/layout/activity_main.xml
@@ -27,12 +27,12 @@
     <com.google.android.gms.maps3d.Map3DView
         android:id="@+id/map3dView"
         map3d:mode="hybrid"
-        map3d:centerLat="40.748392"
-        map3d:centerLng="-73.986060"
-        map3d:centerAlt="175"
-        map3d:heading="26"
-        map3d:tilt="67"
-        map3d:range="4000"
+        map3d:centerLat="39.982129291022446"
+        map3d:centerLng="-105.30156359691158"
+        map3d:centerAlt="2483"
+        map3d:heading="270"
+        map3d:tilt="75"
+        map3d:range="6000"
         map3d:roll="0"
         map3d:minAltitude="0"
         map3d:maxAltitude="1000000"
